+++ slot_link/slot_link_ui_parts.py
@@ -144,7 +144,7 @@ def draw_reset_animation_selector(layout: bpy.types.UILayout, action: bpy.types.
 			row.label(text="The Reset Animation has no Targets!", icon="ERROR")
 
 
-def draw_link_buttons(layout: bpy.types.UILayout, action: bpy.types.Action, only_one_button: bool = False, scale: float = 1):
+def draw_link_buttons(layout: bpy.types.UILayout, action: bpy.types.Action, only_one_button: bool = False, scale: float = 1, vertical_only: bool = False):
 	"""The main 'Link Slots' buttons"""
 	# Prepare legacy/newly created Action
 	if(action.is_action_legacy):
@@ -160,13 +160,13 @@ def draw_link_buttons(layout: bpy.types.UILayout, action: bpy.types.Action, only
 		return
 
 	# Main link button
-	row = layout.row(align=True)
+	row = layout.row(align=True) if not vertical_only else layout.column()
 	row.alignment = "EXPAND"
 	row.alert = not check_action(action)
 	row.scale_x = row.scale_y = scale
 	row.operator(LinkSlots.bl_idname, text="Link Slots", icon="DECORATE_LINKED").use_reset = True
 	if(not only_one_button and action.slot_link.reset_animation):
-		row = row.row(align=True)
+		row = row.row(align=True) if not vertical_only else layout.column()
 		row.alignment = "RIGHT"
 		row.operator(LinkSlots.bl_idname, text="..without Reset").use_reset = False